Implemented net manager functions

+++ src/network/src/ros/src/NetworkContainer.cpp
@@ -9,7 +9,7 @@
 #include <ros/ros.h>
 
 INetworkManager& NetworkContainer::getNetworkManager() {
-    static HashMap<uint16_t, uint16_t, gs_MAX_AGENT_IN_MAP> s_hashMap;
+    static HashMap<uint16_t, uint32_t, gs_MAX_AGENT_IN_MAP> s_hashMap;
     static NetworkManager s_networkManager(LoggerContainer::getLogger(), s_hashMap);
 
     return s_networkManager;
